requirements.txt
# Bibliotecas principais
pandas>=2.0.0
numpy>=1.24.0

# Banco de dados
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Geração de dados
faker>=18.0.0
openpyxl>=3.1.0

# Visualização
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine Learning
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0

# Jupyter
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.25.0